# swagger.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: WorkLogs API
  description: API
  contact:
    email: michal.dering@globallogic.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:5000/v1
paths:
...
  '/worklogs':
    get:
      summary: Get all workLogs
      operationId: getWorkLogs
      tags:
        - worklog
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkLog'
        '401':
          description: Unauthorized Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '404':
          description: WorkLog not found
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultServerError"
    post:
      summary: Add a new workLog
      operationId: addWorkLog
      tags:
        - worklog
      requestBody:
        description: WorkLog object that needs to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkLogToAdd'
      responses:
        '201':
          description: Success
        '401':
          description: Unauthorized Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultServerError"
  '/worklogs/{id}':
    get:
      summary: Get workLog by id
      operationId: getWorkLog
      tags:
        - worklog
      parameters:
        - name: id
          in: path
          description: id of the workLog that needs to be fetched
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkLog'
        #'400':
        #  description: Invalid username supplied
        '401':
          description: Unauthorized Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '404':
          description: WorkLog not found
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultServerError"
    put:
      summary: Update workLog
      operationId: updateWorkLog
      tags:
        - worklog
      parameters:
        - name: id
          in: path
          description: id of the workLog that needs to be updated
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: WorkLog object for update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkLogToUpdate'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkLogUpdated'
        #'400':
        #  description: Invalid user supplied
        '401':
          description: Unauthorized Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '404':
          description: WorkLog not found
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultServerError"
    delete:
      summary: Delete workLog
      operationId: deleteWorkLog
      tags:
        - worklog
      parameters:
        - name: id
          in: path
          description: id of the workLog that needs to be deleted
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
        #'400':
        #  description: Invalid username supplied
        '401':
          description: Unauthorized Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '404':
          description: WorkLog not found
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultServerError"
security:
  - bearerAuth: []           
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UnauthorizedError:
      type: object
      properties:
        status:
          type: string
        info:
          type: object
          properties:
            message:
              type: string
            expiredAt:
              type: integer
              format: int32
            tokenValidFrom:
              type: integer
              format: int32
    DefaultServerError:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        statusCode:
          type: integer
          format: int32
        errorCode:
          type: integer
          format: int32
        result:
          type: array
          items: {}
...
    WorkLog:
      type: object
      properties:
        id:
          type: integer
          format: int32
        userId:
          type: integer
          format: int32
        taskId:
          type: integer
          format: int32
        reportedHours:
          type: integer
          format: int32
    WorkLogToAdd:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        taskId:
          type: integer
          format: int32
        reportedHours:
          type: integer
          format: int32
    WorkLogToUpdate:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        taskId:
          type: integer
          format: int32
        reportedHours:
          type: integer
          format: int32
    WorkLogUpdated:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        taskId:
          type: integer
          format: int32
        reportedHours:
          type: integer
          format: int32